# Backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Anteiku API
  description: TODO
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local HTTP Development Server
paths:
...
  /auth/login:
    post:
      summary: "Send authentication request to a server"
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuthDto'
      responses:
        200:
          description: "Returns access token"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthResponseDto'
        401:
          $ref: '#/components/responses/UnauthorizedError'
  /auth/refresh:
    post:
      summary: "Refreshes the user's authentication tokens or redirects him to the login page"
      operationId: refreshAuthTokens
      responses:
        200:
          description: "Returns cookie with a refreshed authentication token"
          headers:
            set-cookie:
              content:
                text/plain:
                  schema:
                    type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthTokensDto'
        303:
          description: "Refresh token is expired, authentication is required. User will be redirected to the login page"
          headers:
            Location:
              content:
                text/plain:
                  schema:
                    const: "http://0.0.0.0:3000/login" # https://github.com/OAI/OpenAPI-Specification/issues/2512
        401:
          description: "Invalid access token"
...
  /friends:
    get:
      summary: "Get my friends list"
      operationId: getMyFriends
      security:
        - bearerAuth: [ ]
      responses:
        200:
          description: "List of friends"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FriendDto'
  /friends/{userId}:
      post:
        summary: "Send a friend request"
        operationId: addFriend
        parameters:
          - name: userId
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          200:
            description: "Request sent"
          400:
            description: "User not found or already friends"

      put:
        summary: "Accept a friend request"
        operationId: acceptFriend
        parameters:
          - name: userId
            in: path
            required: true
            description: "The ID of the person who sent the request"
            schema:
              type: string
              format: uuid
        responses:
          200:
            description: "Friendship accepted"

      delete:
        summary: "Remove friend or cancel request"
        operationId: removeFriend
        parameters:
          - name: userId
            in: path
            required: true
            schema:
              type: string
              format: uuid
        responses:
          200:
            description: "Removed"
...
  /friends/{userId}/block:
      post:
        summary: "Block a user"
        description: "Blocks the user. If a friendship or request exists, it is overridden."
        operationId: blockUser
        parameters:
          - name: userId
            in: path
            required: true
            description: "ID of the user to block"
            schema:
              type: string
              format: uuid
        responses:
          200:
            description: "User blocked successfully"
          400:
            description: "Cannot block self or user not found"

      delete:
        summary: "Unblock a user"
        operationId: unblockUser
        parameters:
          - name: userId
            in: path
            required: true
            description: "ID of the user to unblock"
            schema:
              type: string
              format: uuid
        responses:
          200:
            description: "User unblocked successfully"
...
components:
  responses:
    UnauthorizedError:
      description: "Access token is missing"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserId:
      type: string
      format: uuid
      description: "User's universally unique identifier"
...
    UserPassword:
      type: string
      format: password
    UserEmail:
      type: string
      format: email
...
    UserAuthDto:
      type: object
      description: "User authentication data"
      properties:
        email:
          $ref: '#/components/schemas/UserEmail'
        password:
          $ref: '#/components/schemas/UserPassword'
    UserAuthTokensDto:
      type: object
      description: "User's access+refresh tokens"
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        refreshTokenExpiresAt:
          type: string
          format: date-time
    UserAuthResponseDto:
      type: object
      description: "User and token info"
      properties:
        authTokens:
          $ref: '#/components/schemas/UserAuthTokensDto'
        userInfo:
          $ref: '#/components/schemas/UserInfoDto'
    UserInfoDto:
      type: object
      description: "User general account info"
      properties:
        id:
          $ref: '#/components/schemas/UserId'
        email:
          $ref: '#/components/schemas/UserEmail'
        username:
          type: string
        role:
          type: string
...
    FriendDto:
      type: object
      description: "Friend relationship informational data"
      properties:
        id:
          $ref: '#/components/schemas/UserId'
        username:
          type: string
        displayName:
          type: string
        picture:
          type: string
        about:
          type: string
        status:
          type: string
          enum: [ONLINE, IDLE, DND, OFFLINE]
        friendStatus:
          type: string
          enum: [FRIEND, PENDING, BLOCKED]
...